Extract a `ui/search_panel` component from `orders/cart`
