+++ admin/app/components/solidus_admin/orders/cart/component.yml
@@ -3,6 +3,3 @@
 en:
   title: "Cart"
   search_placeholder: "Find a variant by name or SKU"
-  loading: "Loading..."
-  initial: "Type to search"
-  empty: "No results"
