--- admin/app/components/solidus_admin/orders/cart/result/component.js
@@ -1,7 +1,0 @@
-import { Controller } from '@hotwired/stimulus'
-
-export default class extends Controller {
-  submit(event) {
-    event.currentTarget.requestSubmit()
-  }
-}
