+++ admin/app/components/solidus_admin/ui/search_panel/component.yml
@@ -1,0 +1,4 @@
+en:
+  loading: "Loading..."
+  initial: "Type to search"
+  empty: "No results"
