+++ admin/spec/components/previews/solidus_admin/ui/search_panel/component_preview.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+# @component "ui/search_panel"
+class SolidusAdmin::UI::SearchPanel::ComponentPreview < ViewComponent::Preview
+  include SolidusAdmin::Preview
+
+  def overview
+    render_with_template
+  end
+end
